feat: add anime and episode detail complement entities and databases
This commit introduces new features for handling anime and episode detail complements.
- Adds `AnimeDetailComplement` and `EpisodeDetailComplement` entities for storing additional data.
- Implements `AnimeDetailComplementConverter` and `EpisodeDetailComplementConverter` for Room type conversion.
- Creates `AnimeDetailComplementDatabase` and `EpisodeDetailComplementDatabase` for local data storage.
- Adds `AnimeDetailComplementDao` and `EpisodeDetailComplementDao` for data access.
- Updates `DatabaseModule` to provide dependencies for the new databases and DAOs.
- Updates `fragment_anime_watch.xml` to use `watching_episode` color instead of `selected_episode` color.
- Updates `StreamingUtils` to use `watching_episode` color instead of `selected_episode` color.
- Updates `colors.xml` to remove `selected_episode` color.
- Updates `IntroOutroHandler` to use long as time type.
- Updates `AnimeDetailDatabase` file name.

+++ app/src/main/java/com/example/animeapp/models/TimeRange.kt
@@ -7,6 +7,6 @@ import kotlinx.serialization.Serializable
 @Parcelize
 @Serializable
 data class TimeRange(
-    val start: Int,
-    val end: Int
+    val start: Long,
+    val end: Long
 ): Parcelable
